sdk/rust: Fix pwrite to flushing buffer before returning
Let's call flush() after write_all() to be safe. It's unclear to me if
this explains a hostfs test failure I saw, but Tokio's documentation
suggests that a write_all() returns when the write is handed over to the
thread pool, but makes no guarantees that the write actually is done,
meaning a subsequent read might return state data.